Move KeyValue, KeyCode and KeyLocation from ext.* to dom.*.
